Extract isElement to own file
