+++ lib/samsam.js
@@ -1,7 +1,7 @@
 "use strict";
 
 var o = Object.prototype;
-var div = typeof document !== "undefined" && document.createElement("div");
+var isElement = require("./is-element");
 
 function isNaN(value) {
     // Unlike global isNaN, this avoids type coercion
@@ -43,26 +43,6 @@ function isArguments(object) {
         return true;
     }
     return false;
-}
-
-/**
- * @name samsam.isElement
- * @param Object object
- *
- * Returns ``true`` if ``object`` is a DOM element node. Unlike
- * Underscore.js/lodash, this function will return ``false`` if ``object``
- * is an *element-like* object, i.e. a regular object with a ``nodeType``
- * property that holds the value ``1``.
- */
-function isElement(object) {
-    if (!object || object.nodeType !== 1 || !div) { return false; }
-    try {
-        object.appendChild(div);
-        object.removeChild(div);
-    } catch (e) {
-        return false;
-    }
-    return true;
 }
 
 /**
